+++ crates/bevy_gizmos_render/Cargo.toml
@@ -30,6 +30,7 @@ bevy_utils = { path = "../bevy_utils", version = "0.20.0-dev" }
 bevy_core_pipeline = { path = "../bevy_core_pipeline", version = "0.20.0-dev" }
 bevy_transform = { path = "../bevy_transform", version = "0.20.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.20.0-dev" }
+bevy_shapes = { path = "../bevy_shapes", version = "0.20.0-dev", features = [ "meshing" ] }
 bevy_color = { path = "../bevy_color", version = "0.20.0-dev" }
 bevy_material = { path = "../bevy_material", version = "0.20.0-dev" }
 bevy_log = { path = "../bevy_log", version = "0.20.0-dev" }
