+++ crates/bevy_light/Cargo.toml
@@ -26,10 +26,12 @@ bevy_color = { path = "../bevy_color", version = "0.20.0-dev", features = [
 ] }
 bevy_gizmos = { path = "../bevy_gizmos", version = "0.20.0-dev", optional = true }
 bevy_log = { path = "../bevy_log", version = "0.20.0-dev" }
+bevy_shapes = { path = "../bevy_shapes", version = "0.20.0-dev", optional = true }
+bevy_gizmos = { path = "../bevy_gizmos", version = "0.20.0-dev", optional = true }
 
 # other
 tracing = { version = "0.1", default-features = false }
-wgpu-types = { version = "29.0.3", default-features = false }
+wgpu-types = { version = "30", default-features = false }
 half = "2.7.1"
 smallvec = { version = "1", default-features = false }
 
@@ -38,6 +40,7 @@ default = []
 experimental_pbr_pcss = []
 webgl = []
 webgpu = []
+bevy_gizmos = ["bevy_shapes?/gizmos", "dep:bevy_gizmos", "dep:bevy_shapes"]
 
 [lints]
 workspace = true
