+++ crates/bevy_pbr/src/render/light.rs
@@ -32,9 +32,8 @@ use bevy_material::{
     MaterialProperties,
 };
 use bevy_math::{
-    ops,
-    primitives::{HalfSpace, ViewFrustum},
-    proj, Mat4, UVec4, Vec3, Vec3Swizzles, Vec4, Vec4Swizzles,
+    camera::{HalfSpace, ViewFrustum},
+    ops, proj, Mat4, UVec4, Vec3, Vec3Swizzles, Vec4, Vec4Swizzles,
 };
 use bevy_mesh::{Mesh3d, MeshVertexBufferLayoutRef};
 use bevy_platform::collections::{HashMap, HashSet};
