+++ crates/bevy_shapes/src/meshing/dim2.rs
@@ -4,13 +4,10 @@ use core::mem;
 use bevy_asset::RenderAssetUsages;
 use bevy_mesh::{Indices, Mesh, MeshBuilder, Meshable, VertexAttributeValues};
 
-use crate::{
-    meshing::{extrusion::PerimeterSegment, Extrudable},
-    primitives::{
-        Annulus, Capsule2d, Circle, CircularSector, CircularSegment, ConvexPolygon, Ellipse,
-        Polyline2d, Primitive2d, Rectangle, RegularPolygon, Rhombus, Ring, Segment2d, Triangle2d,
-        Triangle3d, WindingOrder,
-    },
+use crate::primitives::{
+    Annulus, Capsule2d, Circle, CircularSector, CircularSegment, ConvexPolygon, Ellipse,
+    Polyline2d, Primitive2d, Rectangle, RegularPolygon, Rhombus, Ring, Segment2d, Triangle2d,
+    Triangle3d, WindingOrder,
 };
 use bevy_math::{ops, FloatExt, Vec2, Vec3};
 use bevy_reflect::prelude::*;
@@ -64,16 +61,6 @@ impl MeshBuilder for CircleMeshBuilder {
             .mesh_builder()
             .resolution(self.resolution)
             .build()
-    }
-}
-
-impl Extrudable for CircleMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        alloc::vec![PerimeterSegment::Smooth {
-            first_normal: Vec2::Y,
-            last_normal: Vec2::Y,
-            indices: (0..self.resolution).chain([0]).collect(),
-        }]
     }
 }
 
@@ -215,24 +202,6 @@ impl MeshBuilder for CircularSectorMeshBuilder {
     }
 }
 
-impl Extrudable for CircularSectorMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        let (sin, cos) = ops::sin_cos(self.sector.arc.half_angle);
-        let first_normal = Vec2::new(sin, cos);
-        let last_normal = Vec2::new(-sin, cos);
-        alloc::vec![
-            PerimeterSegment::Flat {
-                indices: alloc::vec![self.resolution, 0, 1],
-            },
-            PerimeterSegment::Smooth {
-                first_normal,
-                last_normal,
-                indices: (1..=self.resolution).collect(),
-            },
-        ]
-    }
-}
-
 impl Meshable for CircularSector {
     type Output = CircularSectorMeshBuilder;
 
@@ -353,24 +322,6 @@ impl MeshBuilder for CircularSegmentMeshBuilder {
     }
 }
 
-impl Extrudable for CircularSegmentMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        let (sin, cos) = ops::sin_cos(self.segment.arc.half_angle);
-        let first_normal = Vec2::new(sin, cos);
-        let last_normal = Vec2::new(-sin, cos);
-        alloc::vec![
-            PerimeterSegment::Flat {
-                indices: alloc::vec![self.resolution, 0, 1],
-            },
-            PerimeterSegment::Smooth {
-                first_normal,
-                last_normal,
-                indices: (1..=self.resolution).collect(),
-            },
-        ]
-    }
-}
-
 impl Meshable for CircularSegment {
     type Output = CircularSegmentMeshBuilder;
 
@@ -441,14 +392,6 @@ impl MeshBuilder for ConvexPolygonMeshBuilder {
     }
 }
 
-impl Extrudable for ConvexPolygonMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        alloc::vec![PerimeterSegment::Flat {
-            indices: (0..self.vertices.len() as u32).chain([0]).collect(),
-        }]
-    }
-}
-
 /// A builder used for creating a [`Mesh`] with a [`RegularPolygon`] shape.
 #[derive(Clone, Copy, Debug, Reflect)]
 #[reflect(Default, Debug, Clone)]
@@ -506,14 +449,6 @@ impl MeshBuilder for RegularPolygonMeshBuilder {
             .mesh_builder()
             .resolution(self.sides)
             .build()
-    }
-}
-
-impl Extrudable for RegularPolygonMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        alloc::vec![PerimeterSegment::Flat {
-            indices: (0..self.sides).chain([0]).collect(),
-        }]
     }
 }
 
@@ -592,16 +527,6 @@ impl MeshBuilder for EllipseMeshBuilder {
         .with_inserted_attribute(Mesh::ATTRIBUTE_NORMAL, normals)
         .with_inserted_attribute(Mesh::ATTRIBUTE_UV_0, uvs)
         .with_inserted_indices(Indices::U32(indices))
-    }
-}
-
-impl Extrudable for EllipseMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        alloc::vec![PerimeterSegment::Smooth {
-            first_normal: Vec2::Y,
-            last_normal: Vec2::Y,
-            indices: (0..self.resolution).chain([0]).collect(),
-        }]
     }
 }
 
@@ -785,24 +710,6 @@ impl MeshBuilder for AnnulusMeshBuilder {
     }
 }
 
-impl Extrudable for AnnulusMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        let vert_count = 2 * self.resolution;
-        alloc::vec![
-            PerimeterSegment::Smooth {
-                first_normal: Vec2::NEG_Y,
-                last_normal: Vec2::NEG_Y,
-                indices: (0..vert_count).step_by(2).chain([0]).rev().collect(), // Inner hole
-            },
-            PerimeterSegment::Smooth {
-                first_normal: Vec2::Y,
-                last_normal: Vec2::Y,
-                indices: (1..vert_count).step_by(2).chain([1]).collect(), // Outer perimeter
-            },
-        ]
-    }
-}
-
 impl Meshable for Annulus {
     type Output = AnnulusMeshBuilder;
 
@@ -876,14 +783,6 @@ impl MeshBuilder for RhombusMeshBuilder {
     }
 }
 
-impl Extrudable for RhombusMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        alloc::vec![PerimeterSegment::Flat {
-            indices: alloc::vec![0, 1, 2, 3, 0],
-        }]
-    }
-}
-
 impl Meshable for Rhombus {
     type Output = RhombusMeshBuilder;
 
@@ -950,21 +849,6 @@ impl MeshBuilder for Triangle2dMeshBuilder {
     }
 }
 
-impl Extrudable for Triangle2dMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        let is_ccw = self.triangle.winding_order() == WindingOrder::CounterClockwise;
-        if is_ccw {
-            alloc::vec![PerimeterSegment::Flat {
-                indices: alloc::vec![0, 1, 2, 0],
-            }]
-        } else {
-            alloc::vec![PerimeterSegment::Flat {
-                indices: alloc::vec![2, 1, 0, 2],
-            }]
-        }
-    }
-}
-
 /// A builder used for creating a [`Mesh`] with a [`Rectangle`] shape.
 #[derive(Clone, Copy, Debug, Reflect)]
 #[reflect(Default, Debug, Clone)]
@@ -999,33 +883,7 @@ impl RectangleMeshBuilder {
 
 impl MeshBuilder for RectangleMeshBuilder {
     fn build(&self) -> Mesh {
-        let [hw, hh] = [self.half_size.x, self.half_size.y];
-        let positions = alloc::vec![
-            [hw, hh, 0.0],
-            [-hw, hh, 0.0],
-            [-hw, -hh, 0.0],
-            [hw, -hh, 0.0],
-        ];
-        let normals = alloc::vec![[0.0, 0.0, 1.0]; 4];
-        let uvs = alloc::vec![[1.0, 0.0], [0.0, 0.0], [0.0, 1.0], [1.0, 1.0]];
-        let indices = Indices::U32(alloc::vec![0, 1, 2, 0, 2, 3]);
-
-        Mesh::new(
-            PrimitiveTopology::TriangleList,
-            RenderAssetUsages::default(),
-        )
-        .with_inserted_indices(indices)
-        .with_inserted_attribute(Mesh::ATTRIBUTE_POSITION, positions)
-        .with_inserted_attribute(Mesh::ATTRIBUTE_NORMAL, normals)
-        .with_inserted_attribute(Mesh::ATTRIBUTE_UV_0, uvs)
-    }
-}
-
-impl Extrudable for RectangleMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        alloc::vec![PerimeterSegment::Flat {
-            indices: alloc::vec![0, 1, 2, 3, 0],
-        }]
+        Mesh::quad_mesh(self.half_size)
     }
 }
 
@@ -1154,32 +1012,6 @@ impl MeshBuilder for Capsule2dMeshBuilder {
         .with_inserted_attribute(Mesh::ATTRIBUTE_NORMAL, normals)
         .with_inserted_attribute(Mesh::ATTRIBUTE_UV_0, uvs)
         .with_inserted_indices(Indices::U32(indices))
-    }
-}
-
-impl Extrudable for Capsule2dMeshBuilder {
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        let resolution = self.resolution;
-        let top_semi_indices = (0..resolution).collect();
-        let bottom_semi_indices = (resolution..(2 * resolution)).collect();
-        alloc::vec![
-            PerimeterSegment::Smooth {
-                first_normal: Vec2::X,
-                last_normal: Vec2::NEG_X,
-                indices: top_semi_indices,
-            }, // Top semi-circle
-            PerimeterSegment::Flat {
-                indices: alloc::vec![resolution - 1, resolution],
-            }, // Left edge
-            PerimeterSegment::Smooth {
-                first_normal: Vec2::NEG_X,
-                last_normal: Vec2::X,
-                indices: bottom_semi_indices,
-            }, // Bottom semi-circle
-            PerimeterSegment::Flat {
-                indices: alloc::vec![2 * resolution - 1, 0],
-            }, // Right edge
-        ]
     }
 }
 
@@ -1370,72 +1202,6 @@ where
         } else {
             panic!("The inner and outer meshes should have the same number of vertices, and have required attributes");
         }
-    }
-}
-
-impl<P> Extrudable for RingMeshBuilder<P>
-where
-    P: Primitive2d + Meshable,
-    P::Output: Extrudable,
-{
-    /// A list of the indices each representing a part of the perimeter of the mesh.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the following assumptions are not met.
-    ///
-    /// It is assumed that the inner and outer meshes have the same number of vertices.
-    /// If not, then the [`MeshBuilder`] of the underlying 2d primitive has generated
-    /// a different number of vertices for the inner and outer instances of the primitive.
-    ///
-    /// It is assumed that the `primitive_topology` of the mesh returned by
-    /// the underlying builder is [`PrimitiveTopology::TriangleList`]
-    /// and that the mesh has [`Mesh::ATTRIBUTE_POSITION`], [`Mesh::ATTRIBUTE_NORMAL`] and [`Mesh::ATTRIBUTE_UV_0`] attributes.
-    fn perimeter(&self) -> Vec<PerimeterSegment> {
-        let outer_vertex_count = self
-            .get_vertex_attributes()
-            .filter(|r| r.outer_positions.len() == r.inner_positions.len())
-            .expect("The inner and outer meshes should have the same number of vertices, and have required attributes")
-            .outer_positions
-            .len();
-
-        let mut outer_perimeter = self.outer_shape_builder.perimeter();
-        let inner_perimeter =
-            self.inner_shape_builder
-                .perimeter()
-                .into_iter()
-                .rev()
-                .map(|segment| match segment {
-                    PerimeterSegment::Smooth {
-                        first_normal,
-                        last_normal,
-                        mut indices,
-                    } => PerimeterSegment::Smooth {
-                        first_normal: -last_normal,
-                        last_normal: -first_normal,
-                        indices: {
-                            let outer_perimeter_vertex_count = outer_vertex_count as u32;
-                            indices.reverse();
-                            for i in &mut indices {
-                                *i += outer_perimeter_vertex_count;
-                            }
-                            indices
-                        },
-                    },
-                    PerimeterSegment::Flat { mut indices } => PerimeterSegment::Flat {
-                        indices: {
-                            let outer_perimeter_vertex_count = outer_vertex_count as u32;
-                            indices.reverse();
-                            for i in &mut indices {
-                                *i += outer_perimeter_vertex_count;
-                            }
-                            indices
-                        },
-                    },
-                });
-
-        outer_perimeter.extend(inner_perimeter);
-        outer_perimeter
     }
 }
