+++ crates/bevy_sprite_render/src/sprite_mesh/mod.rs
@@ -9,8 +9,8 @@ use bevy_ecs::{
 use bevy_asset::{Assets, Handle};
 
 use bevy_image::TextureAtlasLayout;
-use bevy_math::{primitives::Rectangle, vec2};
-use bevy_mesh::{Mesh, Mesh2d, MeshAttributeCompressionFlags, MeshBuilder, Meshable};
+use bevy_math::Vec2;
+use bevy_mesh::{Mesh, Mesh2d, MeshAttributeCompressionFlags};
 
 use bevy_platform::collections::HashMap;
 use bevy_shader::load_shader_library;
@@ -50,9 +50,7 @@ fn add_mesh(
 ) {
     let quad = quad.get_or_insert_with(|| {
         meshes.add(
-            Rectangle::from_size(vec2(1.0, 1.0))
-                .mesh_builder()
-                .build()
+            Mesh::quad_mesh(Vec2::ONE)
                 .with_removed_attribute(Mesh::ATTRIBUTE_NORMAL)
                 .compressed_mesh(
                     MeshAttributeCompressionFlags::COMPRESS_POSITION
